Extract `types.ts`
