--- lib/modules/datasource/github-tags/types.ts
@@ -1,11 +1,0 @@
-export interface TagResponse {
-  object: {
-    type: string;
-    url: string;
-    sha: string;
-  };
-}
-
-export interface GitHubTag {
-  name: string;
-}
